+++ app/models/customer.py
@@ -1,7 +1,8 @@
-from sqlalchemy import Column, String, Text, Boolean, DateTime, Integer
+from sqlalchemy import Column, String, Text, Boolean, DateTime, Integer, ForeignKey
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.core.database import engine, Base, get_db
+
 
 class Customer(Base):
     __tablename__ = "customers"
@@ -24,3 +25,5 @@ class Customer(Base):
     # Future expansion
     # alerts = relationship("Alert", back_populates="customer")
     # users = relationship("User", back_populates="customer")
+    sources = relationship("Source", back_populates="customer")
+    detail = relationship("CustomerDetail", back_populates="customer", cascade="all, delete", uselist=False)